Add files via upload
fixed the problem that in the sexual scenario, its hard for the propagules rains individual find its mate within the same habitat. So the colonization from propagules rains process change to the method that colonizing a pair of female and male individual at one. Note : the pairwise_propagues_rains_num = 1/2*propagues_rains_num. and propagues_rains_num should be a even number.